Reworked the project structure, abstracted functions, renamed element ids
Each different high-level tab has its own layout which can be found inside the respective inside the apps/ directory. Each simply contains a div with the second-level tabs and a div for the content. Tied to each of them is a callback that returns the layout (or "options") for the chosen level-two tab.  These callbacks will probably be merged later in a bigger callback for all the tabs.

These latter layouts can be found in the sub-directories of each tab: one python file per level-two option. The main logic of each level-two tab will go here.

I also added a few functions in the utilities to avoid some repetition. The menus were ported in their own file. Any configuration is ported to the main app.py (might be renamed later to server.py or something). Finally, index.py was modified to fit previous changes. It now is easier to get a sense of the layout of each layer, but to thoroughly understand a specific element you will need to dive deeper than what was necessary before. Still, I believe this structure makes more sense and will save us maintenance troubles in the long run.