Added connection to GoogleSheets, altered api callbacks, slightly modified Vaso's additions
The code in `APIs.py` is ugly and extremely repetitive but when I tried to put them in a list and pass that list in all the same places, it failed. The reason I needed all these buttons there was because of how I organized my callback to update on both `connect_button` clicks and on tab change. If this button disappears from the layout then the callback complains. The reason the list might not be working may have something to do with https://github.com/plotly/dash-renderer/issues/23 or I could be screwing somewhere with unpacking.